+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * CONRAT-42439: Migrate unit tests from Quick/Nimble to Testing framework
 * CONRAT-42855: Update deprecated URLSession initializer for unit test mock objects
 * CONRAT-43326: Re-enable SwiftLint.
+* CONRAT-43324: Remove NSObject.ratTracker from AnalyticsManager and use RAnalyticsRATTracker shared instance instead.
 
 ### Build fixes
 * CONRAT-43264: fix the Swift Package Test Target.